+++ src/main/java/roidrole/patternbanners/recipe/HEIPlugin.java
@@ -9,38 +9,33 @@ import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.BannerPattern;
-import net.minecraftforge.common.config.ConfigCategory;
 import net.minecraftforge.oredict.OreDictionary;
+import roidrole.patternbanners.Tags;
 import roidrole.patternbanners.config.ConfigGeneral;
-import roidrole.patternbanners.config.ConfigMapping;
 import roidrole.patternbanners.recipe.wrapper.PatternApplyWrapper;
 import roidrole.patternbanners.recipe.wrapper.PatternFromShapeWrapper;
-import roidrole.patternbanners.recipe.wrapper.RecipeAddPatternWrapper;
-
-import java.util.Collections;
+import roidrole.patternbanners.recipe.wrapper.PatternOnlyShapeWrapper;
 
 import static roidrole.patternbanners.loom._Loom.loomBlock;
-import static roidrole.patternbanners.recipe._Recipe.categoryUid;
+import static roidrole.patternbanners.recipe._Recipe.PATTERNS_ONLY_SHAPE;
+import static roidrole.patternbanners.recipe._Recipe.PATTERN_APPLY_RECIPES;
 
 @JEIPlugin
 public class HEIPlugin implements IModPlugin {
+    public static String categoryUid = Tags.MOD_ID+"pattern";
     @Override
     public void register(IModRegistry registry){
 
         registry.handleRecipes(PatternApply.class, PatternApplyWrapper::new, categoryUid);
-
-        for(ConfigCategory mapping : ConfigMapping.mappings){
-            registry.addRecipes(Collections.singleton(new PatternApply(mapping)), categoryUid);
-        }
+        registry.addRecipes(PATTERN_APPLY_RECIPES, categoryUid);
 
         if(ConfigGeneral.shapes_pattern){
             registry.handleRecipes(PatternFromShape.class, PatternFromShapeWrapper::new, VanillaRecipeCategoryUid.CRAFTING);
         } else {
-            for(BannerPattern pattern : BannerPattern.values()){
-                if(pattern.hasPatternItem() || !pattern.hasPattern()){continue;}
-                registry.addRecipes(Collections.singletonList(new RecipeAddPatternWrapper(pattern)), categoryUid);
-            }
+            registry.handleRecipes(BannerPattern.class, PatternOnlyShapeWrapper::new, categoryUid);
+            registry.addRecipes(PATTERNS_ONLY_SHAPE, categoryUid);
         }
+
 
         registry.addRecipeCatalyst(new ItemStack(Items.BANNER, 1, OreDictionary.WILDCARD_VALUE), categoryUid);
         if(ConfigGeneral.Recipes.craftingTable) {